Extract reusable panorama reconstruction pipeline
